% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_mat.R
\name{A_mat}
\alias{A_mat}
\title{Additive relationship matrix from pedigree}
\usage{
A_mat(ped, ploidy, order.ped = TRUE, inverse = FALSE)
}
\arguments{
\item{ped}{Pedigree in 3 or 7 column format (see Details)}

\item{ploidy}{integer or NULL for mixed ploidy pedigree}

\item{order.ped}{TRUE/FALSE does the pedigree need to be ordered so that progeny follow parents}

\item{inverse}{TRUE/FALSE to return matrix inverse}
}
\value{
Additive relationship matrix (dim: indiv x indiv) or its inverse
}
\description{
Additive relationship matrix from pedigree
}
\details{
Founders should have NA for mother and father. Any parents not listed in the first column are added as founders. 

This is a wrapper that prepares the pedigree in the format required for R package \code{polyAinv} by Hamilton and Kerr (cite them if you use this function). A random bivalents model for tetraploid meiosis is assumed.

Pedigree assumes mother (dam) before father (sire). 7-col format pedigree is needed for mixed ploidy: id, dam, sire, dam.gamete, dam.lambda, sire.gamete, sire.lambda (Hamilton and Kerr 2018).
}
\references{
Hamilton and Kerr (2018) Theor Appl Genet 131:851-860 doi:10.1007/s00122-017-3041-y
}
