% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADsplit.R
\name{ADsplit}
\alias{ADsplit}
\title{Extract read counts from AD string}
\usage{
ADsplit(AD, ALT, n.core = 1)
}
\arguments{
\item{AD}{array of AD strings}

\item{ALT}{TRUE or FALSE (= REF)}

\item{n.core}{number of cores}
}
\value{
integer data with same dimensions as AD
}
\description{
Extract read counts from AD string
}
