% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2VCF.R
\name{array2VCF}
\alias{array2VCF}
\title{Generate pseudo-VCF file from SNP array data}
\usage{
array2VCF(VCF.file, ploidy, AD, GT = NULL, map = NULL, header = NULL)
}
\arguments{
\item{VCF.file}{name of VCF file to create}

\item{ploidy}{ploidy}

\item{AD}{matrix of allele depths, see \code{\link{readXY}}}

\item{GT}{optional, matrix of genotype dosages (0,1,2..ploidy)}

\item{map}{optional, 3 column data frame (marker,chrom,pos)}

\item{header}{optional, header text for the VCF file}
}
\description{
Creates VCF file with GT and AD fields
}
\details{
AD and GT matrices should be markers x indiv. Per the VCF standard, the "." is used for missing data.
}
