% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ploidy.R
\name{check_ploidy}
\alias{check_ploidy}
\title{Check ploidy}
\usage{
check_ploidy(geno, map)
}
\arguments{
\item{geno}{Genotype matrix (markers x indiv)}

\item{map}{Data frame with marker map (Marker, Chrom, Position)}
}
\value{
List containing
\describe{
\item{mat}{Matrix (indiv x chrom) of results}
\item{plot}{ggplot2 barplot}
}
}
\description{
Fraction of simplex or triplex markers
}
\details{
For every indiv in the genotype matrix, the fraction of markers per chromosome called as simplex or triplex is calculated, which should be low for diploids. A small amount of missing genotype data can be tolerated.
}
