% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trio.R
\name{check_trio}
\alias{check_trio}
\title{Check markers for parent-offspring trio}
\usage{
check_trio(parentage, geno, ploidy)
}
\arguments{
\item{parentage}{Data frame with three columns: id, mother, father}

\item{geno}{Matrix of allele dosages: markers x indiv}

\item{ploidy}{2 or 4}
}
\value{
Data frame with the percentage of incompatible markers for each trio
}
\description{
Check markers for parent-offspring trio
}
\details{
Computes the percentage of markers at which the two parents and offspring have incompatible allele dosages (for tetraploids, the random bivalents model is used). For dihaploid offspring of a single tetraploid parent, use \code{ploidy} = 4 and "haploid" for the father in \code{parentage}, as well as a diploid (0,1,2) genotype for the offspring. A small amount of missing genotype data can be tolerated.
}
