% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Genetic contribution}
\usage{
contribution(id, ped)
}
\arguments{
\item{id}{vector of target names}

\item{ped}{pedigree}
}
\value{
matrix of contributions, with rows \code{id} and columns for all members of \code{ped}
}
\description{
Computes genetic contribution from pedigree
 
Argument \code{ped} is three column pedigree; can be created using \code{\link{get_pedigree}}.
}
