% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2PO.R
\name{dart2PO}
\alias{dart2PO}
\title{Convert DArTag data to PolyOrigin format}
\usage{
dart2PO(
  infile,
  outfile,
  first.data.row = 9,
  first.data.col = 6,
  array.file = NULL
)
}
\arguments{
\item{infile}{input filename}

\item{outfile}{output filename}

\item{first.data.row}{first data row}

\item{first.data.col}{first data column}

\item{array.file}{optional}
}
\value{
List of two data frames with statistics for markers and samples
}
\description{
Convert DArTag data to PolyOrigin format
}
\details{
Designed for standard two-row format from DArT. Column 1 contains the AlleleID in format MarkerName|Haplotype. Haplotypes are named Ref,RefMatch,Alt,AltMatch,Other. Counts are combined for Ref + RefMatch, as well as Alt + AltMatch. Other haplotypes are discarded. Genotype calls can using the updog package (Gerard et al. 2018). If a sample has no reads at a marker, the genotype is NA and posterior probability equals 0.
}
