% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr2vcf.R
\name{fr2vcf}
\alias{fr2vcf}
\title{Convert Genome Studio Final Report to VCF}
\usage{
fr2vcf(fr.file, map.file, vcf.file)
}
\arguments{
\item{fr.file}{name of Genome Studio Final Report file}

\item{map.file}{vcf file with map positions for the markers}

\item{vcf.file}{output vcf file}
}
\description{
Convert Genome Studio Final Report to VCF
}
\details{
XY values are multiplied by 100 to generate AD field
}
