% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_coeff.R
\name{G_coeff}
\alias{G_coeff}
\title{Coefficients to make the additive G matrix}
\usage{
G_coeff(geno, ploidy)
}
\arguments{
\item{geno}{Matrix of allele dosages: markers x indiv}

\item{ploidy}{2 or 4}
}
\value{
List containing
\describe{
\item{coeff}{Coefficients of the marker effects (dim: indiv x marker)}
\item{scale}{Scaling factor between markers and indiv}
}
}
\description{
Coefficients of the additive effects for bi-allelic markers
}
\details{
Additive effects are based on the traditional orthogonal decomposition of genetic variance in panmictic populations (Fisher 1918; Kempthorne 1957; Endelman et al. 2018). The G matrix can be computed from the coefficients and scaling factor according to G = tcrossprod(coeff/scale). Missing genotype data is replaced with the population mean.
}
\references{
Fisher (1918) Trans. Roy. Soc. Edin. 52:399-433.

Kempthorne (1957) An Introduction to Genetic Statistics.

Endelman et al. (2018) Genetics 209:77-87.
}
