% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.R
\name{gbs}
\alias{gbs}
\title{Genotype calls for GBS}
\usage{
gbs(in.file, out.file, ploidy, bias = TRUE, n.core = 1, silent = FALSE)
}
\arguments{
\item{in.file}{VCF input file}

\item{out.file}{VCF output file}

\item{ploidy}{ploidy}

\item{bias}{TRUE/FALSE, whether to estimate allelic bias}

\item{n.core}{number of cores}

\item{silent}{TRUE/FALSE}
}
\value{
nothing
}
\description{
Genotype calls for genotype-by-sequencing (GBS) data
}
\details{
VCF input file must contain AD field. Posterior mode and mean genotypes are added as GT and DS fields. GQ is also added based on probability of posterior mode. Binomial calculation uses R/updog package (Gerard et al. 2018) with "norm" prior. Previous INFO is discarded; adds NS, DP.AVG, AF.GT, AB, OD, SE.
}
