% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs_call.R
\name{gbs_call}
\alias{gbs_call}
\title{Genotype calls from allele count file}
\usage{
gbs_call(in.file, out.base, ploidy, n.core = 1)
}
\arguments{
\item{in.file}{input file}

\item{out.base}{base name for output files}

\item{ploidy}{ploidy}

\item{n.core}{number of cores}
}
\value{
marker x indiv matrix of read depths
}
\description{
Genotype calls from allele count file
}
\details{
Function uses R/updog package (Gerard et al. 2018), which is based on a binomial model. If a sample has no reads at a marker, the genotype is NA and posterior probability equals 0. 

The input file follows the format generated by \code{\link{dart_tag}}, with the read counts for each marker x id combination represented as "count1|count2". 

Three output files are generated with the (1) allele ratio, (2) posterior mode, and (3) posterior mean.
}
