% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_call.R
\name{geno_call}
\alias{geno_call}
\title{Genotype calls}
\usage{
geno_call(
  data,
  filename,
  model.ploidy = 4L,
  sample.ploidy = 4L,
  min.posterior = 0,
  transform = TRUE
)
}
\arguments{
\item{data}{matrix (markers x id) of input values for the normal mixture model}

\item{filename}{CSV filename with the model parameters}

\item{model.ploidy}{2 or 4 (default)}

\item{sample.ploidy}{2 or 4 (default)}

\item{min.posterior}{minimum posterior probability (default 0) for genotype call}

\item{transform}{TRUE (default) or FALSE whether to apply arcsin square root transformation}
}
\value{
matrix of allele dosages (0,1,2,..ploidy) with dimensions markers x individuals
}
\description{
Genotype calls based on a normal mixture model
}
\details{
The first column of the CSV input file should be the SNP ID, followed by columns for the normal distribution means, standard deviations, and mixture probabilities. Genotype calls are based on the maximum a posteriori (MAP) method. If the posterior probability of the MAP genotype is less than \code{min.posterior}, then \code{NA} is returned for that sample. By default, an arcsin square root transformation is applied to the input values to match the approach used by R package fitPoly. To use a tetraploid mixture model for diploid samples, set \code{sample.ploidy = 2} and \code{model.ploidy = 4}.
}
