% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parents.R
\name{get_parents}
\alias{get_parents}
\title{Get parents of genotypes}
\usage{
get_parents(id, pedfile, delim = ",", na.string = "NA", DH = FALSE)
}
\arguments{
\item{id}{Vector of target names}

\item{pedfile}{Name of pedigree file}

\item{delim}{Delimiter for the pedigree file (default is "," for CSV)}

\item{na.string}{String used for NA in the pedigree file (default is "NA")}

\item{DH}{TRUE/FALSE should 4x parent of dihaploid be returned}
}
\value{
Data frame with columns id, mother, father
}
\description{
Get parents of genotypes
}
\details{
The function generates data frame with the parents of the input \code{id}. For lines with any degree of inbreeding (more than one en-dash) or dihaploids (-DH), no look-up table is needed. Otherwise, the parents are searched in the three-column \code{pedfile} (id,mother,father),  The id column can be the identifier for an individual or cross. String matches must be exact or based on the naming convention crossID-progenyID.
}
