% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pedigree.R
\name{get_pedigree}
\alias{get_pedigree}
\title{Generate pedigree}
\usage{
get_pedigree(id, pedfile, delim = ",", na.string = "NA", trim = TRUE)
}
\arguments{
\item{id}{Vector of names of individuals}

\item{pedfile}{Name of pedigree file}

\item{delim}{Delimiter for the pedigree file (default is "," for CSV)}

\item{na.string}{String used for NA in the pedigree file (default is "NA")}

\item{trim}{TRUE/FALSE whether to trim pedigree (see Details)}
}
\value{
Data frame with columns id, mother, father
}
\description{
Generate pedigree for a set of individuals
}
\details{
Finds ancestors of individuals in a three-column pedigree file (id,mother,father). The id column can be the identifier for an individual or cross. String matches must be exact or based on the naming convention crossID-progenyID. The returned pedigree is ordered using R package \code{pedigree} so that offspring follow parents. When \code{trim} is TRUE (default), the pedigree is trimmed to remove ancestors with only one offspring (which are not needed to compute the pedigree relationship matrix).
}
