% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GvsA.R
\name{GvsA}
\alias{GvsA}
\title{Plot G vs. A}
\usage{
GvsA(
  parentage,
  G,
  A,
  filename = NULL,
  thresh.G = Inf,
  thresh.A = 0.5,
  Gmax = NULL,
  Amax = NULL
)
}
\arguments{
\item{parentage}{Data frame of individuals to plot, with 3 columns: id,mother,father}

\item{G}{Genomic relationship matrix}

\item{A}{Pedigree relationship matrix}

\item{filename}{Name of PDF file to save the results (optional for one individual)}

\item{thresh.G}{Threshold above which names are displayed (default Inf)}

\item{thresh.A}{Threshold above which names are displayed (default 0.5)}

\item{Gmax}{Upper limit for y-axis for plotting. If NULL, maximum value in G is used.}

\item{Amax}{Upper limit for x-axis for plotting. If NULL, maximum value in A is used.}
}
\description{
Plot marker-based vs. pedigree-based additive relationship coefficients
}
\details{
Useful for finding and correcting pedigree errors. If the G or A coefficient for an individual exceeds the threshold, its name is displayed in the figure. If parentage contains one individual, by default a ggplot2 variable will be returned, but the result can also be written to file. If multiple individuals are present, a filename is required.
}
