% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing marker data}
\usage{
impute(geno, method)
}
\arguments{
\item{geno}{Matrix of allele dosages with dimensions markers x indiv}

\item{method}{Either "mean" or "mode"}
}
\value{
Imputed genotype matrix (markers x indiv)
}
\description{
Impute marker data based on the population mean or mode
}
\details{
Missing values are imputed with either the population mean or mode (most frequent value) for each marker
}
