% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternal_lineage.R
\name{maternal_lineage}
\alias{maternal_lineage}
\title{Generate pedigree}
\usage{
maternal_lineage(id, pedfile, delim = ",", na.string = "NA")
}
\arguments{
\item{id}{Vector of names of individuals}

\item{pedfile}{Name of pedigree file}

\item{delim}{Delimiter for the pedigree file (default is "," for CSV)}

\item{na.string}{String used for NA in the pedigree file (default is "NA")}
}
\value{
Data frame with columns id, lineage
}
\description{
Generate pedigree for a set of individuals
}
\details{
Finds ancestors of individuals in a three-column pedigree file (id,mother,father). The id column can be the identifier for an individual or cross. String matches must be exact or based on the naming convention crossID-progenyID. The returned pedigree is ordered using R package \code{pedigree} so that offspring follow parents. When \code{trim} is TRUE (default), the pedigree is trimmed to remove ancestors with only one offspring (which are not needed to compute the pedigree relationship matrix).
}
