% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MME.R
\docType{class}
\name{MME-class}
\alias{MME-class}
\alias{MME}
\title{S4 class for solving the mixed model equations}
\description{
S4 class for solving the mixed model equations
}
\section{Slots}{

\describe{
\item{\code{y}}{response}

\item{\code{X}}{design matrix for the fixed effects}

\item{\code{Z}}{design matrix for random genetic effects. Colnames must match rownames of the matrices in \code{K}.}

\item{\code{kernels}}{list of variance-covariance matrices for the genetic effects. Matrices must have the same rownames attribute.}

\item{\code{Rmat}}{residual variance-covariance matrix}
}}

