% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MME.R
\name{MME}
\alias{MME}
\title{Solve the Mixed Model Equations}
\usage{
MME(y, X, Z, Vu, Rmat)
}
\arguments{
\item{y}{response (vector)}

\item{X}{design matrix for the fixed effects}

\item{Z}{list of design matrices for the random effects}

\item{Vu}{list of covariance matrices for the random effects}

\item{Rmat}{residual variance-covariance matrix}
}
\value{
List containing
\describe{
\item{blue}{BLUE for the average fixed effect (numeric)}
\item{blup}{list of data frames, one for each random effect, with two variables: BLUP, r2}
}
}
\description{
Solve the Mixed Model Equations
}
\details{
Details
}
\keyword{internal}
