% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_NMM.R
\name{predict_NMM}
\alias{predict_NMM}
\title{Predict allele dosage}
\usage{
predict_NMM(x, params, ploidy, min.posterior = 0, transform = TRUE)
}
\arguments{
\item{x}{vector of input values}

\item{params}{vector of parameters for the normal mixture model}

\item{ploidy}{2 or 4}

\item{min.posterior}{minimum posterior probability for a genotype call to be made}

\item{transform}{TRUE/FALSE whether to apply an arcsin square root transformation}
}
\value{
vector of allele dosages (0,1,2,..ploidy)
}
\description{
Returns allele dosage as the maximum a posteriori (MAP) genotype based on a normal mixture model (NMM)
}
\details{
The input values should fall within the interval [0,1]. The vector of model parameters should be ordered as follows: means, standard deviations, mixture probabilities. If the posterior probability of the MAP genotype is less than `min.posterior`, then `NA` is returned for that sample. By default, an arcsin square root transformation is applied to the input values, as this is the approach used by the `fitPoly` package.
}
\keyword{internal}
