% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{Stage1}
\alias{Stage1}
\title{Stage 1 analysis of multi-environment trials}
\usage{
Stage1(
  data,
  traits,
  fixed = NULL,
  random = NULL,
  silent = TRUE,
  workspace = "500mb",
  pworkspace = "500mb"
)
}
\arguments{
\item{data}{Data frame with phenotype data}

\item{traits}{Vector of column names from \code{data}}

\item{fixed}{Vector of column names from \code{data}}

\item{random}{Vector of column names from \code{data}}

\item{silent}{TRUE/FALSE, whether to suppress ASReml-R output}

\item{workspace}{Memory limit for ASRreml-R variance estimation}

\item{pworkspace}{Memory limit for ASRreml-R BLUE computation}
}
\value{
List containing
\describe{
\item{aic}{AIC from ASReml-R}
\item{blue}{data frame of BLUEs}
\item{vcov}{variance-covariance matrix of the BLUEs}
\item{h2}{matrix with heritability, genetic, and residual correlations (see Details)}
}
}
\description{
Stage 1 analysis of multi-environment trials
}
\details{
Stage 1 of the two-stage approach described by Damesa et al. 2017, using ASReml-R for variance component estimation (license is required). The variable \code{data} must have a column labeled "id" with the names of the different genotypes (i.e., clones or individuals). To include other variables (besides "id") in the model, include them in \code{fixed} or \code{random} as appropriate, and make sure they have the correct type in the data frame: factor vs. numeric. If multiple traits are included, a multivariate analysis is performed, and only plots with data for all traits are included. The \code{h2} matrix returned by the function contains the estimated genetic correlations above the diagonal, residual correlations below the diagonal, and plot-based heritability on the diagonal. For multivariate analysis, the data frame \code{blue} returned by the function is in long format, with a column named "trait". By default, the workspace and pworkspace limits for ASReml-R are set at 500mb. If you get an error about insufficient memory, try increasing the appropriate value (workspace for variance estimation and pworkspace for BLUE computation).
}
\references{
Damesa et al. 2017. Agronomy Journal 109: 845-857. doi:10.2134/agronj2016.07.0395
}
