% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step2.R
\name{Step2}
\alias{Step2}
\title{Genomic prediction of marker effects from multi-environment trials}
\usage{
Step2(pheno, G, D = NULL, AIC = TRUE, silent = FALSE, workspace = "128mb")
}
\arguments{
\item{pheno}{List of matrices containing BLUEs and variance-covariance matrix}

\item{G}{Returned object from \code{\link{G_mat}} for additive effects}

\item{D}{Optional, returned object from \code{\link{D_mat}} for dominance effects (default is NULL)}

\item{AIC}{Boolean variable, whether to select A vs. AD model based on AIC (default is TRUE)}

\item{silent}{Boolean variable, whether to suppress ASReml-R convergence monitoring (default is FALSE)}

\item{workspace}{Workspace memory for ASReml-R (default is "128mb")}
}
\value{
List containing
\describe{
\item{params}{Matrix with AIC and estimated parameters}
\item{indiv}{Matrix of predicted values for the individuals in \code{geno}}
\item{markers}{Matrix of predicted effects for the markers in \code{geno}}
}
}
\description{
Additive and dominance marker effects predicted by BLUP, based on variance components estimated with ASReml-R (must have license)
}
\details{
Best practice for the analysis of datasets comprising multiple experiments follows a two-step approach (Damesa et al. 2017). Step 1 generates a vector of genotype BLUEs (and variance-covariance matrix) for each experiment, taking its design and potentially spatial factors into account. In Step 2, the BLUEs are used as a response variable in a mixed model, and the covariance matrix for the residuals is constrained to equal the direct sum of the covariance matrices of the BLUEs. In the current implementation, the Step 2 model contains three genetic effects: additive, digenic dominant, and independent (to capture higher-order non-additive effects). The digenic dominant term has nonzero mean to allow for heterosis/inbreeding depression (Varona et al. 2018). The model also contains a fixed effect for each experiment and independent random effects for the genotype x environment interaction (an environment may contain multiple experiments). Each element of \code{pheno} corresponds to an experiment and is a list with components named "env" and "y". Env is the name of the environment for that experiment, and "y" is a matrix of dimensions n x (n+1), where n is the number of genotypes in the trial. The first column of the matrix contains the BLUEs, and the remaining n columns are the variance-covariance matrix for the BLUEs. ASReml-R is used to estimate variance components, and BLUPs are computed based on standard formulas (Searle et al. 1992). When \code{D} is not NULL, variances are first estimated for the strictly additive (A) model and used as initial values when fitting the additive + dominance (AD) model. If \code{AIC = TRUE}, then whichever model (A vs. AD) has lower AIC is used for BLUP; otherwise, BLUPs are computed based on the AD model regardless of the AIC values. Individuals in \code{pheno} but not in \code{G} are removed from the analysis; to make predictions for unphenotyped individuals, include them in \code{G} (and \code{D}).
}
\references{
Damesa et al. 2017. Agronomy Journal 109: 845-857. doi:10.2134/agronj2016.07.0395

Varona et al. 2018. Frontiers in Genetics 9: 78. doi:10.3389/fgene.2018.00078

Searle et al. 1992. Variance Components. doi:10.1002/9780470316856
}
