% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_ped.R
\name{trim_ped}
\alias{trim_ped}
\title{Trim pedigree}
\usage{
trim_ped(ped, focal)
}
\arguments{
\item{ped}{three column pedigree: id, parent1, parent2}

\item{focal}{vector of id's for the focal population}
}
\value{
trimmed pedigree
}
\description{
Trim pedigree
}
\details{
Removes founders with a single child, provided the founder is not in the focal population.
}
\keyword{internal}
