% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alias.R
\name{update_alias}
\alias{update_alias}
\title{Update names based on alias}
\usage{
update_alias(x, alias, remove.space = TRUE)
}
\arguments{
\item{x}{Vector of names to update}

\item{alias}{Data frame with two columns: first is the preferred name and second is the alias}

\item{remove.space}{TRUE/FALSE}
}
\value{
Vector with updated names
}
\description{
Update names based on data frame with alias and preferred name
}
\details{
Parameter \code{remove.space} indicates whether blank spaces should be removed before string matching
}
