% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2csv.R
\name{vcf2csv}
\alias{vcf2csv}
\title{Convert VCF to CSV}
\usage{
vcf2csv(vcf.file, csv.file, format)
}
\arguments{
\item{vcf.file}{Input file}

\item{csv.file}{Output file}

\item{format}{Name of FORMAT key to export, either "GT" or "DS"}
}
\value{
none
}
\description{
Convert VCF to CSV
}
