% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2vcf.R
\name{wide2vcf}
\alias{wide2vcf}
\title{Convert Genome Studio XY wide format to VCF}
\usage{
wide2vcf(input.file, map.file, model.file = NULL, ploidy, vcf.file)
}
\arguments{
\item{input.file}{name of input file}

\item{map.file}{vcf file with map positions for the markers}

\item{model.file}{normal mixture model parameters for genotype calls}

\item{ploidy}{sample ploidy, for use with \code{model.file}}

\item{vcf.file}{output vcf file}
}
\description{
Convert Genome Studio XY wide format to VCF
}
\details{
XY values are multiplied by 100 to generate AD field.
}
