% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_mat.R
\name{A_mat}
\alias{A_mat}
\title{Additive relationship matrix from pedigree}
\usage{
A_mat(ped, ploidy, order.ped = TRUE)
}
\arguments{
\item{ped}{Pedigree in three column format: id, mother, father}

\item{ploidy}{2 or 4}

\item{order.ped}{TRUE/FALSE does the pedigree need to be ordered so that progeny follow parents}
}
\value{
Additive relationship matrix (dim: indiv x indiv)
}
\description{
Additive relationship matrix from pedigree
}
\details{
This is a wrapper that prepares the pedigree in the format required for R package \code{AGHmatrix} by Amadeu et al. (2016) (cite them if you use this function). A random bivalents model for tetraploid meiosis is assumed.
}
\references{
Amadeu et al. (2016) Plant Genome 9, doi:10.3835/plantgenome2016.01.0009
}
