% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_tag.R
\name{dart_tag}
\alias{dart_tag}
\title{Extract Ref/Alt counts from DArTag data file}
\usage{
dart_tag(filename)
}
\arguments{
\item{filename}{input filename}
}
\value{
3D array of allele counts with dimensions: markers, samples, alleles (ref/alt)
}
\description{
Extract Ref/Alt counts from DArTag data file
}
\details{
Designed for standard two-row format from DArT. First 11 rows contain sample information. Column 1 contains the AlleleID in format MarkerName|Haplotype. Haplotypes are named Ref,RefMatch,Alt,AltMatch,Other. Counts are combined for Ref + RefMatch, as well as Alt + AltMatch. Other haplotypes are discarded.
}
