% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_tag.R
\name{dart_tag}
\alias{dart_tag}
\title{Convert DArTag file to read count file}
\usage{
dart_tag(
  dart.file,
  first.data.row = 9,
  first.data.col = 6,
  out.file,
  map.file = NULL,
  AB.file = NULL
)
}
\arguments{
\item{dart.file}{DArTag CSV filename}

\item{first.data.row}{first data row in dart.file}

\item{first.data.col}{first data column in dart.file}

\item{out.file}{name of output file}

\item{map.file}{optional CSV file (marker, chrom, position) to integrate into the output}

\item{AB.file}{optional CSV file (marker, REF) to convert to allele B dosage}
}
\value{
marker x indiv matrix of read depths
}
\description{
Convert DArTag file to read count file
}
\details{
Designed for standard two-row format from DArT. Column 1 contains the AlleleID in format MarkerName|Haplotype. Haplotypes are named Ref,RefMatch,Alt,AltMatch,Other. Counts are combined for Ref + RefMatch, as well as Alt + AltMatch. Other haplotypes are discarded. 

Output format contains the read counts for each marker x id combination in the format "count1|count2" (which is the input format for PolyOrigin).

Use \code{AB.file} to convert REF/ALT counts from dart.file to A/B counts from SNP array. The file must have columns named "marker" and "REF", where REF is either A or B. Only markers present in AB.file will be in the output.
}
