% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_mat.R
\name{G_mat}
\alias{G_mat}
\title{Genomic relationship matrix}
\usage{
G_mat(geno, ploidy, p.ref = NULL, method = "VR1", sep = "_", n.core = 1)
}
\arguments{
\item{geno}{genotype matrix or filename}

\item{ploidy}{ploidy}

\item{p.ref}{optional, reference population frequency for method "VR1"}

\item{method}{"VR1" or "AM"}

\item{sep}{character separating id from haplotype for "AM" method}

\item{n.core}{number of cores, only used with "AM"}
}
\value{
G matrix
}
\description{
Genomic relationship matrix
}
\details{
For \code{method}="VR1", Method 1 of VanRaden (2008) is used, and its polyploid extension (Endelman et al. 2018). Missing data is replaced with the population mean for each marker. If \code{p.ref} is NULL, the current population is used as the reference population. For "VR1", \code{geno} is an allele dosage matrix for bi-allelic loci (sites x indiv). 

For \code{method}="AM", the Allele Matching coefficient is calculated, which is the probability that two haplotypes sampled at random (with replacement) are identical (Weir and Goudet 2017). Missing data are not allowed. For "AM", \code{geno} is a phased genotype matrix, with alleles coded as positive integers. The name of each column is the id and haplotype concatenated with a separating character, \code{sep}. This character needs to be unique (i.e., not present in id or haplotype).
}
\references{
VanRaden (2008) J. Dairy Sci 91:4414-4423.

Endelman et al. (2018) Genetics 209:77-87.

Weir and Goudet (2017) Genetics 206:2085-2103.
}
