% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_mat.R
\name{G_mat}
\alias{G_mat}
\title{Additive genomic relationships}
\usage{
G_mat(geno, ploidy)
}
\arguments{
\item{geno}{Matrix of allele dosages (markers x indiv)}

\item{ploidy}{Any even integer (2,4,6,...)}
}
\value{
G matrix
}
\description{
Relationship matrix for additive effects with bi-allelic markers
}
\details{
Additive effects are based on the traditional orthogonal decomposition of genetic variance in panmictic populations (Fisher 1918; Kempthorne 1957; Endelman et al. 2018). Missing genotype data is replaced with the population mean.
}
\references{
Fisher (1918) Trans. Roy. Soc. Edin. 52:399-433.

Kempthorne (1957) An Introduction to Genetic Statistics.

Endelman et al. (2018) Genetics 209:77-87.
}
