% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_L2H.R
\name{impute_L2H}
\alias{impute_L2H}
\title{Impute from low to high density markers by Random Forest}
\usage{
impute_L2H(
  high.file,
  low.file,
  out.file = NULL,
  params = list(),
  exclude = NULL,
  n.core = 1
)
}
\arguments{
\item{high.file}{name of high density file}

\item{low.file}{name of low density file}

\item{out.file}{name of CSV output file for imputed data}

\item{params}{list of parameters (see Details)}

\item{exclude}{optional, vector of high density samples to exclude}

\item{n.core}{multicore processing}
}
\value{
matrix of OOB error with dimensions markers x trees. For regression model, it is MSE.
}
\description{
Impute from low to high density markers by Random Forest
}
\details{
Argument \code{params} is a list with the following options: format, model, n.tree, n.mark. \code{format} can have values "GT" (integer dosage) or "DS" (real numbers between 0 and ploidy). \code{model} can be "class" for classification or "regress" for regression when "GT" is used; for "DS" format, only regression is permitted. \code{n.tree} is the number of trees (default = 100). \code{n.mark} is the number of markers to use as predictors (default = 100), chosen based on minimum distance to the target. 

The \code{exclude} argument is useful for cross-validation. 

Both VCF and CSV are allowable input file formats--they are recognized based on the file extension. For CSV, the first three columns should be marker, chrom, pos. The output file is CSV. 

Any missing data are imputed separately for each input file at the outset, using the population mean (regress) or mode (class) for each marker.
}
