% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_L2H.R
\name{impute_L2H}
\alias{impute_L2H}
\title{Impute from low to high density markers}
\usage{
impute_L2H(
  high.file,
  low.file,
  out.file,
  ploidy,
  params = list(format = "GT", n.tree = 100, n.mark = 100),
  n.core = 1
)
}
\arguments{
\item{high.file}{name of high density VCF file}

\item{low.file}{name of low density VCF file}

\item{out.file}{name of CSV output file for the imputed dataset}

\item{ploidy}{ploidy}

\item{params}{list of parameters (see Details)}

\item{n.core}{multicore processing}
}
\value{
vector of mean OOB error (vs. number of trees)
}
\description{
Impute from low to high density markers
}
\details{
Missing data in high density file is imputed with the population mean (DS) or mode (GT) for each marker. The Random Forest package is used for imputing up to high density. Argument \code{params} recognizes the following parameters: "format","n.tree","n.mark". Format can have values "GT" or "DS" (default = "GT"), n.tree is the number of trees (default = 100), and n.mark is the number of markers to use as predictors (default = 200), chosen based on minimum distance to the target. Classification trees are used for GT and regression trees for DS. Any markers in low.file that are not present in the high.file or with an identical position are discarded.
}
