% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_PO.R
\name{impute_PO}
\alias{impute_PO}
\title{Impute from low to high density markers with PolyOrigin}
\usage{
impute_PO(
  high.file,
  low.file,
  low.format = "GT",
  ped.file,
  out.file,
  posterior = "mean",
  exclude = NULL
)
}
\arguments{
\item{high.file}{name of high density file with phased parents}

\item{low.file}{name of low density VCF file with progeny}

\item{low.format}{either "GT" (default) or "AD"}

\item{ped.file}{pedigree file for progeny (must follow PO format)}

\item{out.file}{name of CSV output file for imputed data}

\item{posterior}{either "max" or "mean" (default)}

\item{exclude}{optional, vector of high density samples to exclude}
}
\description{
Impute from low to high density markers by linkage analysis with PolyOrigin
}
\details{
You must have separately installed PolyOrigin and Julia for this function to work.

#' The high density file contains phased parental genotypes in PolyOrigin format. The first 3 columns are the genetic map in cM: marker, chrom, position. To output imputed data with physical rather than genetic map positions, including a fourth column named "bp". Subsequent columns are the phased parental genotypes. 

VCF is assumed for the low-density file. The pedigree file must follow PolyOrigin format.
}
