% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madc.R
\name{madc}
\alias{madc}
\title{Multi-Allelic Haplotype Counts from DArTag}
\usage{
madc(madc.file, marker = "CDF1")
}
\arguments{
\item{madc.file}{MADC filename}

\item{marker}{Name of marker ("CDF1" is only option so far)}
}
\value{
matrix of haplotype counts
}
\description{
Multi-Allelic Haplotype Counts from the DArTag MADC (Missing Allele Discovery Count) file
}
\details{
Due to multi-allelism, for some trait markers a correct interpretation is not possible using the collapsed counts file; the MADC file is needed. Currently, the only marker implemented is CDF1 for potato DArTag. The CDF1 marker detects the 2C, 2T, and 4 alleles, and all other haplotypes are treated as allele 1. Allele 3 is not detected by the assay.
}
