% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madc.R
\name{madc}
\alias{madc}
\title{Multi-allelic markers}
\usage{
madc(madc.file, marker, min.AD = 5)
}
\arguments{
\item{madc.file}{MADC file}

\item{marker}{Name of marker}
}
\value{
data frame of counts and diploid GT
}
\description{
Multi-allelic markers
}
