% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{Stage1}
\alias{Stage1}
\title{Stage 1 analysis of multi-environment trials}
\usage{
Stage1(
  data,
  traits,
  effects = NULL,
  silent = TRUE,
  workspace = "500mb",
  pworkspace = "500mb"
)
}
\arguments{
\item{data}{data frame with phenotype data}

\item{traits}{vector of column names from \code{data}}

\item{effects}{list of other effects in the model}

\item{silent}{TRUE/FALSE, whether to suppress ASReml-R output}

\item{workspace}{memory limit for ASRreml-R variance estimation}

\item{pworkspace}{memory limit for ASRreml-R BLUE computation}
}
\value{
List containing
\describe{
\item{H2}{matrix of broad-sense heritability for each env x trait combination}
\item{blue}{data frame of BLUEs for id x traits}
\item{blue.vcov}{list of BLUE + variance-covariance matrices (one matrix per trait)}
}
}
\description{
Stage 1 analysis of multi-environment trials
}
\details{
Stage 1 of the two-stage approach described by Damesa et al. 2017, using ASReml-R for variance component estimation (license is required). The variable \code{data} must have one column labeled "id" for the individuals,  one labeled "env" for the environments, plus columns for each of the traits to be analyzed. The data for each environment x trait combination are analyzed independently with a linear mixed model. Argument \code{effects} is a named list of character vectors to specify other effects in the model. Each vector has two elements: the first is "fixed" or "random", and the second is "factor" or "numeric". For example, to include a random block effect, use \code{effects=list(block=c("random","factor"))}. To include stand.count as a numeric covariate, use \code{effects=list(stand.count=c("fixed","numeric"))}. By default, the workspace and pworkspace limits for ASReml-R are set at 500mb. If you get an error about insufficient memory, try increasing the appropriate value (workspace for variance estimation and pworkspace for BLUE computation).
}
\references{
Damesa et al. 2017. Agronomy Journal 109: 845-857. doi:10.2134/agronj2016.07.0395
}
