% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2_prep.R
\name{Stage2_prep}
\alias{Stage2_prep}
\title{Prepare data for Stage 2 analysis of multi-environment trials}
\usage{
Stage2_prep(blue.vcov, exclude.id = character(0), exclude.env = character(0))
}
\arguments{
\item{blue.vcov}{named list of blue.vcov matrices from Stage 1}

\item{exclude.id}{vector of individuals to exclude}

\item{exclude.env}{vector of envs to exclude}
}
\value{
a list containing
\describe{
\item{blue}{data frame of BLUEs}
\item{Omega}{variance-covariance matrix of BLUEs}
}
}
\description{
Prepare data for Stage 2 analysis of multi-environment trials
}
\details{
Designed to prepare data files for \code{\link{Stage2}} based on output from \code{\link{Stage1}}. Each element of \code{blue.vcov} is a matrix for one trait, with the first column containing BLUEs and the remainder is their variance-covariance. The rownames are the id and env concatenated with ":". The output \code{blue} is a data frame with id, env, and trait (when multiple traits are provided). This allows for multi-trait analysis in Stage 2.
}
